requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
pkg-resources==0.0.0
python-dotenv==0.15.0
requests==2.25.0
urllib3==1.26.2
